requirements.txt
torch
numpy
pandas
datetime
newsapi-python
transformers
plotly
gradio
